Added KDocs and consolidated repeal rule under the rules and amendments paths
