+++ src/main/kotlin/nomic/api/GamesEndpoint.kt
@@ -11,17 +11,34 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
+/**
+ * This controller listens on `api/game` and works to provide, remove, or manipulate data related to games in the database
+ *
+ * @see[nomic.domain.games.GameDomain]
+ * @param[gameDomain] This dependency is a domain service used by the endpoint to manipulate data in the database related to games.
+ */
 @RestController
 @RequestMapping("api/game")
 class GamesEndpoint(val gameDomain: GameDomain) {
-    // Path to endpoint is api/game/create
+
+    /**
+     * This endpoint listens on `api/game/create` and takes in data about a game that will be inserted into the game table in the DB
+     *
+     * @param[input] The game data to eventually insert
+     * @return A spring entity representing the response that gets serialized into JSON
+     */
     @PostMapping("create")
     fun createGame(@RequestBody input: GameModel): ResponseEntity<Any> {
         gameDomain.createGame(input)
         return ResponseEntity("Game Created", HttpStatus.CREATED)
     }
 
-    // Path to endpoint is api/game/remove/ExistingGameId
+    /**
+     * This endpoint listens on `api/game/remove/ExistingGameId` and takes in a gameId that will be deleted from the DB
+     *
+     * @param[gameId] The game id to be deleted from the database
+     * @return A spring entity representing the response that gets serialized into JSON
+     */
     @DeleteMapping("remove/{gameid}", produces = ["application/json;charset=UTF-8"])
     fun deleteGame(@PathVariable(value = "gameid") gameId: String): ResponseEntity<Any> {
         gameDomain.deleteGame(gameId)
