+++ src/main/kotlin/nomic/data/repositories/rulesamendments/RuleAmendmentRepository.kt
@@ -1,15 +1,11 @@
 package nomic.data.repositories.rulesamendments
 
+import nomic.data.EntityNotFoundException
 import nomic.data.dtos.Amendments
 import nomic.data.dtos.Rules
 import nomic.domain.entities.RulesAmendmentsModel
 import org.ktorm.database.Database
-import org.ktorm.dsl.eq
-import org.ktorm.dsl.forEach
-import org.ktorm.dsl.from
-import org.ktorm.dsl.leftJoin
-import org.ktorm.dsl.select
-import org.ktorm.dsl.where
+import org.ktorm.dsl.*
 import org.springframework.stereotype.Repository
 
 /**
@@ -46,5 +42,24 @@ class RuleAmendmentRepository(private val db: Database) : IRuleAmendmentReposito
             }
 
         return rules
+    }
+
+    override fun repealRule(ruleId: Int) {
+        db.update(Amendments) {
+            set(it.active, false)
+            where {
+                it.ruleId eq ruleId
+            }
+        }
+        val result = db.update(Rules) {
+            set(it.active, false)
+            where {
+                it.ruleId eq ruleId
+            }
+        }
+
+        if (result < 1) {
+            throw EntityNotFoundException(ruleId)
+        }
     }
 }
