+++ src/main/kotlin/nomic/domain/rulesamendments/RuleAmendmentDomain.kt
@@ -3,6 +3,7 @@ package nomic.domain.rulesamendments
 import nomic.api.models.AmendmentModel
 import nomic.api.models.RulesAmendmentsApiModel
 import nomic.data.repositories.rulesamendments.RuleAmendmentRepository
+import nomic.domain.entities.RepealRuleResponse
 import nomic.domain.entities.RulesAmendmentsModel
 import org.springframework.stereotype.Service
 
@@ -38,7 +39,7 @@ class RuleAmendmentDomain(
                 )
             }
             if (row.amendId != null && row.ruleActive && row.amendActive!!) {
-                rules.last().amendments?.add(
+                rules.last().amendments.add(
                     AmendmentModel(
                         row.amendId,
                         row.amendIndex!!,
@@ -50,5 +51,12 @@ class RuleAmendmentDomain(
         }
 
         return rules
+    }
+
+    override fun repealRule(ruleId: String): RepealRuleResponse {
+        val ruleIdInt: Int = ruleId.toIntOrNull() ?: throw IllegalArgumentException("Please enter a valid ruleId!")
+        ruleAmendmentRepository.repealRule(ruleIdInt)
+
+        return RepealRuleResponse(true, "Updated Successfully", ruleIdInt)
     }
 }
